requirements.txt
fastapi
uvicorn
python-chess
requests
pydantic
